# Conflicts:
#	aaa-automation-tests/src/test/java/aaa/modules/regression/billing_and_payments/auto_ss/functional/TestRefundProcess.java
#	aaa-automation-tests/src/test/java/aaa/modules/regression/billing_and_payments/helpers/RefundProcessHelper.java
#	aaa-automation-tests/src/test/java/aaa/modules/regression/billing_and_payments/home_ss/ho3/functional/TestRefundProcess.java
#	aaa-automation-tests/src/test/java/aaa/modules/regression/billing_and_payments/pup/functional/TestRefundProcess.java
#	aaa-automation-tests/src/test/java/aaa/modules/regression/sales/auto_ss/functional/TestEValueDiscount.java
#	aaa-automation-tests/src/test/java/aaa/modules/regression/service/helper/HelperCommon.java
#	aaa-automation-tests/src/test/java/aaa/modules/regression/service/home_ss/ho3/functional/TestServiceRFI.java
#	aaa-automation-ui/src/main/java/aaa/common/Tab.java
#	aaa-automation-ui/src/main/java/aaa/main/enums/DocGenEnum.java
#	aaa-automation-ui/src/main/java/aaa/main/metadata/policy/AutoCaMetaData.java
#	aaa-automation-ui/src/main/java/aaa/main/modules/policy/auto_ss/AutoSSPolicy.java

+++ aaa-automation-tests/src/test/java/aaa/helpers/openl/model/home_ca/ho3/HomeCaHO3OpenLPolicy.java
@@ -20,8 +20,8 @@ public class HomeCaHO3OpenLPolicy extends HomeCaOpenLPolicy {
 	@ExcelTableElement(sheetName = OpenLFile.COVERAGE_SHEET_NAME, headerRowIndex = OpenLFile.COVERAGE_HEADER_ROW_NUMBER)
 	private List<HomeCaOpenLCoverage> coverages;
 
-	private Integer covALimit;
-	private Integer covELimit;
+	private Double covALimit;
+	private Double covELimit;
 	private Boolean hasEmployeeDiscount;
 	private Boolean hasMultiPolicyDiscount;
 	private Boolean hasPolicySupportingForm;
@@ -51,19 +51,19 @@ public class HomeCaHO3OpenLPolicy extends HomeCaOpenLPolicy {
 		this.coverages = new ArrayList<>(coverages);
 	}
 
-	public Integer getCovALimit() {
+	public Double getCovALimit() {
 		return covALimit;
 	}
 
-	public void setCovALimit(Integer covALimit) {
+	public void setCovALimit(Double covALimit) {
 		this.covALimit = covALimit;
 	}
 
-	public Integer getCovELimit() {
+	public Double getCovELimit() {
 		return covELimit;
 	}
 
-	public void setCovELimit(Integer covELimit) {
+	public void setCovELimit(Double covELimit) {
 		this.covELimit = covELimit;
 	}
